refactor(codegen): extract index lowering helpers
